requirements.txt
# Core dependencies
Pillow>=10.0.0
numpy>=1.24.0
RPi.GPIO>=0.7.1; sys_platform == "linux"

# Development and build tools
pyinstaller>=5.13.0

# Optional dependencies for enhanced functionality
matplotlib>=3.7.0  # For additional plotting features
opencv-python>=4.8.0  # For advanced image processing
scipy>=1.10.0  # For scientific computing

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
mock>=5.1.0

# Code quality
flake8>=6.0.0
black>=23.0.0
pylint>=2.17.0
